DiveDB/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "DiveDB"
version = "0.1.0"
dependencies = [
  "black",
  "bs4",
  "dash",
  "deltalake",
  "django",
  "djangorestframework",
  "django-storages",
  "duckdb",
  "edfio",
  "flake8",
  "importlib-metadata==4.8.3",
  "load_dotenv",
  "mne",
  "notion_client",
  "pandas",
  "pre-commit",
  "psycopg2-binary",
  "pyarrow",
  "pyedflib",
  "pylint-django",
  "pytest",
  "pytest-django",
  "pytest-recording",
  "python-keystoneclient",
  "python-swiftclient",
  "setuptools",
  "tqdm",
  "xarray",
]

[tool.black]
line-length = 88
target-version = ["py37"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | \.lib
  | _build
  | buck-out
  | build
  | dist
  | services/duck_pond.py
  | pyologger
)/
'''
